security/services/ca/requirements.txt
nose==1.3.0
thrift==0.9.1
PyOpenSSL==0.13.1
pycrypto==2.6.1
kazoo~=1.3.0
ezbake-base-thrift>=2.1
ezsecurity-services>=2.1
ezbake-configuration>=2.1
ezbake-discovery>=2.1
ezbake-thrift-utils>=2.1
